Added mime_utils for mime type validation and conversion, replacing map lookup in http_response.
The implementation is naive but simple. Prioritise the map lookup as is currently done. If the map
lookup fails, check if the provided content type looks valid, and use that if it does. Otherwise
use text/plain. This is then used whenever a mime lookup is performed, rather than only in `set_static_file_info_unsafe`.

See my previous commit message for the loose definition of "valid" that I've applied - checking
it consists of a valid parent type, a slash, and at least one additional character. If a stronger
or weaker definition is preferred, I'm happy to look into that. As the mime type format is available
as a grammar, the additional dependency of CTRE might be handy, but that might also be overkill.

A trivial approach of doing this is to look up against {"valid/", "parent/", "types/"}, with a length check
followed by a strncmp. Strncmp is more c-like than some might be comfortable with, and if a C++17 migration
happens, a std::string_view approach would be nicer. An alternative implementation
would be std::string::substr, but that would introduce up to N string copies for N valid mime parent types.
I'm happy to go with that approach if it's closer to the in-house style.

The addition of a new header file was not an easy decision. I considered adding functions into
mime_types.h courtesy of nginx_mime2cpp.py, but implementing logic as part of codegen is often
undesirable. I considered adding the functions as static functions of http_response, but it would
then add bloat to a core class. I'm happy to revert to one of those approaches if the addition of
a new file is controversial.