Redirect to url with trailing slash, WITHOUT the protocol and host.
Crow does not know the correct protocol when behind a reverse proxy.
So instead, simply redirect to the same url and add a trailing slash.

Note, it is now valid for the "Location" header field to be relative.
From wikipedia:
An obsolete version of the HTTP 1.1 specifications (IETF RFC 2616)
required a complete absolute URI for redirection.
The IETF HTTP working group found that the most popular web browsers
tolerate the passing of a relative URL[3] and, consequently, the updated
HTTP 1.1 specifications (IETF RFC 7231) relaxed the original constraint,
allowing the use of relative URLs in Location headers.