include CPM via CMAKE_CURRENT_SOURCE_DIR
This again allows crow to be used as a CMake subproject.
This broke in #1066, as CMAKE_SOURCE_DIR refers to the directory of the
root CMakeLists.txt. And when being used from a toplevel project,
${CMAKE_SOURCE_DIR}/cmake/CPM.cmake would not refer to crow's copy of CPM.
This change makes the reference relative to the location of the
CMakeLists.txt requiring the include.